refactor: split processing_service.py (1942 lines) into 4 focused modules
