+++ src/services/__init__.py
@@ -8,7 +8,7 @@ from .file_service import FileService
 from .transcription_service import TranscriptionService
 from .speechmatics_service import SpeechmaticsService
 from .enhanced_llm_service import EnhancedLLMService
-from .processing_service import ProcessingService
+from .processing import ProcessingService
 from .base_processing_service import BaseProcessingService
 
 __all__ = [
