+++ src/services/processing/__init__.py
@@ -1,0 +1,4 @@
+"""Processing service package."""
+from .processing_service import ProcessingService
+
+__all__ = ["ProcessingService"]
